+++ contrib/scripts/check-crates.sh
@@ -19,6 +19,7 @@ buildargs=(
     "-p nostr-http-file-storage"
     "-p nostr-database"
     "-p nostr-database-test-suite"
+    "-p nostr-memory"
     "-p nostr-gossip"
     "-p nostr-gossip-memory"
     "-p nostr-gossip-sqlite"