+++ crates/nostr-relay-builder/examples/hyper.rs
@@ -114,7 +114,7 @@ impl Service<Request<Incoming>> for HttpServer {
 async fn main() -> nostr_relay_builder::prelude::Result<()> {
     tracing_subscriber::fmt::init();
 
-    let relay = LocalRelay::new();
+    let relay = LocalRelay::new()?;
 
     let http_addr: SocketAddr = "127.0.0.1:8000".parse()?;
     let listener = TcpListener::bind(&http_addr).await?;