+++ sdk/src/relay/api/send_event.rs
@@ -193,7 +193,7 @@ mod tests {
         let opts = LocalRelayBuilderNip42 {
             mode: LocalRelayBuilderNip42Mode::Write,
         };
-        let mock = LocalRelay::builder().nip42(opts).build();
+        let mock = LocalRelay::builder().nip42(opts).build().unwrap();
         mock.run().await.unwrap();
         let url = mock.url().await;
 
@@ -242,7 +242,7 @@ mod tests {
         let opts = LocalRelayBuilderNip42 {
             mode: LocalRelayBuilderNip42Mode::Write,
         };
-        let mock = LocalRelay::builder().nip42(opts).build();
+        let mock = LocalRelay::builder().nip42(opts).build().unwrap();
         mock.run().await.unwrap();
         let url = mock.url().await;