+++ crates/nostr-relay-builder/CHANGELOG.md
@@ -31,6 +31,7 @@
 - Allow only a single `write` and `query` policy (https://github.com/rust-nostr/nostr/pull/1165)
 - Replace `PolicyResult` with `WritePolicyResult` and `QueryPolicyResult` (https://github.com/rust-nostr/nostr/pull/1166)
 - Remove `tor` feature (https://github.com/rust-nostr/nostr/pull/1253)
+- Change `LocalRelay::new` and `LocalRelayBuilder::build` signatures (https://github.com/rust-nostr/nostr/pull/1263)
 
 ### Changes