+++ database/nostr-database/CHANGELOG.md
@@ -29,6 +29,7 @@
 
 - Remove `Backend::is_persistent` method (https://github.com/rust-nostr/nostr/pull/1138)
 - Use `NonZeroUsize` for bounded capacities instead of `usize` (https://github.com/rust-nostr/nostr/pull/1139)
+- Move `MemoryDatabase` to `nostr-memory` crate (https://github.com/rust-nostr/nostr/pull/1263)
 
 ### Changed